requirements.txt
Pillow==7.2.0
youtube-dl==2020.9.20